Factored out generic AdjointLayoutPlot
